mlxsw: spectrum_qdisc: Unify graft validation
Qdisc graft operations have so far been reported at PRIO, ETS and RED, with RED
events ignored, because RED was not considered a classful qdisc. A following
patch will make mlxsw recognize RED and TBF as classful qdiscs, and thus it is
necessary to validate grafting at these qdiscs as well. Rename the existing
graft validator to make it clear that it is a generic function, and invoke for
RED and TBF graft events as well. Drop the unnecessary PRIO helper and invoke
the graft validator directly for PRIO as well.

Signed-off-by: Petr Machata <petrm@mellanox.com>