mlxsw: spectrum_qdisc: Destroy children in mlxsw_sp_qdisc_destroy()
Currently ETS is the only offloaded classful qdisc, so it handles children
destruction itself. But now it is possible to do it generically for any classful
qdisc. Therefore promote the recursive destruction from the ETS handler to
mlxsw_sp_qdisc_destroy(), so that RED and ETF pick it up in follow-up
patches.

Signed-off-by: Petr Machata <petrm@mellanox.com>