mlxsw: spectrum_qdisc: Clean stats recursively when priomap changes
On Spectrum, there are no per-TC TX counters. Instead, mlxsw uses per-prio
counters and aggregates them according to the priomap. Therefore when priomap
changes, the counter base values need to be reset to reflect the change.
Previously, this was only done for the sole child qdisc, but a following
patch makes RED and ETF classful. Thus apply the request to the whole
sub-tree.

Signed-off-by: Petr Machata <petrm@mellanox.com>