mlxsw: spectrum_router_xm: Introduce basic XM cache flushing
Upon route insertion and removal, it is needed to flush possibly cached
entries from the XM cache. Extend XM op context to carry information
needed for the flush. Implement the flush in delayed work since for HW
design reasons there is a need to wait 50usec before the flush can be
done. If during this time comes the same flush request, consolidate it
to the first one. Implement this queued flushes by a hashtable.

Signed-off-by: Jiri Pirko <jiri@nvidia.com>
---
v1->v2:
- passing u32 instead of pointer to mlxsw_sp_router_ll_xm_mindex_get4()
- s/LSB/MSB/
- added WARN_ON_ONCE() to mlxsw_sp_router_ll_xm_mindex_get6()
- fixed "biggest" comment typo
- remove pointless checks from the start of mlxsw_sp_router_xm_ml_entry_del()
- added rollback to mlxsw_sp_router_xm_ml_entries_add()
- added L-table update order comment
- fixed typo in patch description
- changed some of the comments
- wrapped some over-80-col lines
- moved cache flush schedule after prefixes are inserted/removed
- moved to RLPMCE for all-flush mode instead of XMDR.bulk_entry