net: dev: Make it possible for enqueue egress traffic to backlog [xxx]
Currently the traffic that is enqueued to the backlog by netif_rx() is
always injected to ingress through a call to __netif_receive_skb().

xxxx split to two patches?
xxxx unify want_egress / want_ingress
xxxx describe the race etc. (this should be in the act_mirred patch if split)