selftests: mlxsw: qos_lib: Add a wrapper for running mlnx_qos
mlnx_qos is a script for one-off configuration of DCB. Despite the name it
is not actually Mellanox-specific in any way. It is currently the only
one-off configuration tool available. However, it is very verbose and
parsing out error messages is not really possible. Add a wrapper that makes
it easier to use the tool.

Signed-off-by: Petr Machata <petrm@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>