ethtool: Extend link modes settings uAPI with number of lanes
Currently, when autoneg is on, the user can advertise all the link modes
which correspond to a specific speed. A similar selector for the number
of lanes does not exist. This is significant when a specific speed can
be achieved using different number of lanes. For example, 100Gbps can be
achieved using 2 lanes of 50Gbps or 4 lanes of 25Gbps.

Add 'ETHTOOL_A_LINKMODES_LANES' attribute and expand 'struct
ethtool_link_settings' with lanes field in order to implement a new
lanes selector that will allow the user to advertise link modes with a
specific number of lanes.

When autoneg is off, lanes parameter can be forced only if the driver
supports it. To that end, add a capability bit in 'struct ethtool_ops'
that allows ethtool to know if the driver can handle the lanes parameter
when autoneg is off, so if it does not, an error message is returned.

Example:

$ ethtool -s swp1 lanes 4
$ ethtool swp1
  Settings for swp1:
	Supported ports: [ FIBRE ]
        Supported link modes:   1000baseKX/Full
                                10000baseKR/Full
                                40000baseCR4/Full
				40000baseSR4/Full
				40000baseLR4/Full
                                25000baseCR/Full
                                25000baseSR/Full
				50000baseCR2/Full
                                100000baseSR4/Full
				100000baseCR4/Full
        Supported pause frame use: Symmetric Receive-only
        Supports auto-negotiation: Yes
        Supported FEC modes: Not reported
        Advertised link modes:  40000baseCR4/Full
				40000baseSR4/Full
				40000baseLR4/Full
                                100000baseSR4/Full
				100000baseCR4/Full
        Advertised pause frame use: No
        Advertised auto-negotiation: Yes
        Advertised FEC modes: Not reported
        Speed: Unknown!
        Duplex: Unknown! (255)
        Auto-negotiation: on
        Port: Direct Attach Copper
        PHYAD: 0
        Transceiver: internal
        Link detected: no

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>