ethtool: Add a bitmask for using when autoneg is off
Currently, when autoneg is on, an advertising bitmask is created by
translating the link mode settings from ethtool. That bitmask is then
passed to drivers so that they will advertise the corresponding link
modes.

When autoneg is off, each setting is passed to drivers as is and the
translation to link modes is done in the drivers.

Relieve drivers from the need to perform this translation by passing
them a bitmask of link modes that they need to force, in a similar
fashion to the autoneg on case.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>