mlxsw: ethtool: Add support for forcing number of lanes
Up until now, when autoneg was off, a link modes bitmask corresponding
to the passed attributes (e.g., speed, duplex) was compiled by the
driver.

Now, when ethtool passes such a bitmask to the driver, all that remains
to do is to force that bitmask.

Force the bitmask instead of creating it and set the
'ETHTOOL_CAP_FORCE_LINK_LANES_SUPPORTED' capability bit to indicate to
ethtool that driver can handle forcing of number of lanes.

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>