net: selftests: Add lanes setting test
Test that setting lanes parameter is working.

First, test that when setting only speed, max width is chosen as the
number of lanes.

Second, set max speed and max lanes in the list of advertised link modes,
and then try to set max speed with the lanes below max lanes if exists
in the list.

And last, test that setting number of lanes larger than max lanes fails.

Do the above for both autoneg on and off.

$ ./ethtool_lanes.sh

TEST: When no lanes setting, max width is set                       [ OK ]
TEST: 4 lanes is autonegotiated                                     [ OK ]
TEST: Lanes number larger then max_width is not set                 [ OK ]
TEST: Autoneg off, 4 lanes detected during force mode               [ OK ]
TEST: Lanes number larger then max width is not set                 [ OK ]

Signed-off-by: Danielle Ratson <danieller@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>