mlxsw: reg: Add Switch Port VLAN Stacking Register
SPVTR register configures the VLAN mode of the port to enable VLAN
stacking.

It will be used to configure VXLAN to push VLAN to the decapsulated packet.
Without this setting, Spectrum-2 overtakes the VLAN tag of decapsulated
packet for bridging and derives the VLAN to forward on from the VNI->FID
mapping.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>