mlxsw: reg: Add IRIF related fields for SVFA register
SVFA register controls the VID to FID mapping and {Port, VID} to FID
mapping for virtualized ports.

As preparation for new firmware model, which is called ubridge1, add
some required fields for future use.

On ingress, after ingress ACL, a packet needs to be classified to a FID.
The key for this lookup can be one of:
1. VID. When port is not in virtual mode
2. {RQ, VID}. When port is in virtual mode
3. FID. When FID was set by ingress ACL.

Since RITR no longer performs ingress configuration, the ingress RIF for
the first two entry types need to be set via new fields in SVFA -
'irif_v' and 'irif'.

Add the two mentioned fields for future use.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>