mlxsw: reg: Add VID related fields to SFD
SFD register configures FDB table.
As preparation for new firmware model, which is called ubridge1, add
some required fields for future use.

In the new model, firmware no longer configures the egress VID, this
responsibility is moved to software. For layer 2 this means that software
needs to determine the egress VID for both unicast and multicast.

For unicast FDB records and unicast LAG FDB records, the VID needs to be
set via new fields in SFD - "set_vid" and "vid".

Add the two mentioned fields for future use.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>