mlxsw: reg: Add Switch Multicast Port to Egress VID Register
The SMPE register maps {egress_port, SMPE} -> VID.

The device includes two main tables to support layer 2 multicast (i.e.,
MDB and flooding). These are the PGT (Port Group Table) array and the
MPE (Multicast Port Egress) table. Logically, the pipeline looks as
follows:

(Local port = LP)
     +------------------+             +--------------------------+
     | Port Group Table |{LP list,    |   Multicast Port Egress  | \
MID  |------------------| SMPE index} |--------------------------|  |
---> | LPs bitmap #1    |-----------> | LP #1 eVID .. LP #N eVID |  | SMPE
     | ..               |             | ..                       |  |
     | LPs bitmap #N    |             | ..                       | /
     +------------------+             +--------------------------+
                                       \________________________/
                                                Local port

In Spectrum-1, the index into the MPE table - called switch multicast to
port egress VID (SMPE) - is derived from the PGT entry, whereas in
Spectrum-2 and later ASICs it is derived from the FID.

In the legacy model, software did not interact with this table as it was
completely hidden in firmware. In the new model, software needs to
populate the table itself in order to map from {Local port, SMPE} to an
egress VID. This is done using the SMPE register.

Add the register for future use.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>