mlxsw: reg: Add Router Egress Interface to VID Register
The REIV maps {eRIF, egress_port} -> vid.
As preparation for new firmware model, which is called ubridge1, add
REIV register for future use.

In the past, firmware would take care of the above mentioned mapping,
but in the new model it needs to be performed by software using REIV
register.

The register is reserved when ubridge = 0.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>

 Please enter the commit message for your changes. Lines starting