mlxsw: reg: Add egress FID field to RITR
RITR is used to configure the router interface table.
As preparation for new firmware model, which is called ubridge1,
add egress FID field to RITR.

After routing, a packet needs perform an L2 lookup using the DMAC it got
from the routing and a FID.
In the new model, the egress FID is configured by RITR for both sub-port
and FID RIFs.

Add "efid" field to sub-port router interface and update the comment in
FID router interface.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>