mlxsw: Do not store local port as part of Rx listener
The structure 'mlxsw_listener' is used to run the appropriate function
for processing received packets, according to - local port, trap ID and
mirror reason. Currently the field local port is always set to
'MLXSW_PORT_DONT_CARE', which means that it is ignored. It was probably
added as a preparation for future listeners per port, but it is not
really used. Remove this field.

A next patch will store Rx listeners in a hash-table, this change makes
it easier.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>