mlxsw: pci: Store Rx listeners in hash table
When Rx packet is received, the appropriate Rx listener is searched as part
of core code and then rxl->func() is called to handle the packet.
Currently, Rx listener items are stored in a linked list, which means
that for each received packet, it might take o(n) complexity to find the
listener. This is inefficient and can be improved by storing the items
in a hash table. The key will be {mirror_reason, trap ID} as they are used
today to compare listeners.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>