mlxsw: pci: Prepare Rx buffers for XDP support
To support XDP, an additional headroom should be reserved in the buffers
which are posted to hardware. XDP program is attached per port, but mlxsw
driver does not have a Rx queue per port, so the configurations of XDP
cannot be applied per queue. Reserve XDP_PACKET_HEADROOM in front of all Rx
buffers, which means that when a XDP program is configured, there is no
required change related to queues.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>