mlxsw: pci: Adjust Tx completion poll method for XDP handling
XDP completions are different from SKB completions, they should not be
handled when budget is zero, and XDP API is used to free the frame.

Adjust the existing code to check SDQ number and handle completions
accordingly. Take SKB related code out to an helper function, note that
unmapping the pages is not relevant for XDP as the pages were allocated
from page pool and should be mapped during the lifetime of the pool.

Signed-off-by: Amit Cohen <amcohen@nvidia.com>
Reviewed-by: Ido Schimmel <idosch@nvidia.com>
Signed-off-by: Ido Schimmel <idosch@nvidia.com>